implemented duration unit and limit handling, added new feature inheritPriority, for activities to inherit the priority of the previous activity some small fixes as well

+++ core/src/workflow/org/ofbiz/core/workflow/WfExecutionObject.java
@@ -169,14 +169,14 @@ public interface WfExecutionObject {
      * @throws WfException General workflow exception.
      * @return Getter Priority of
      */
-    public int priority() throws WfException;
+    public long priority() throws WfException;
 
     /**
      * Setter for attribute 'priority'.
      * @param newValue
      * @throws WfException General workflow exception
      */
-    public void setPriority(int newValue) throws WfException;
+    public void setPriority(long newValue) throws WfException;
 
     /**
      * Resume this process or activity.
